backend/requirements.txt
# -------------------------------
# Flask Core & Extensions
# -------------------------------
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0

# -------------------------------
# Database & ORM
# -------------------------------
pymongo==4.6.0

# -------------------------------
# Background Scheduling
# -------------------------------
APScheduler==3.10.4

# -------------------------------
# AI & Translation Services
# -------------------------------
openai==1.51.0
deep-translator==1.11.4
langdetect==1.0.9

# -------------------------------
# Utilities & Security
# -------------------------------
python-dotenv==1.0.0
bcrypt==4.0.1

# -------------------------------
# NLP Support (for TranslationService)
# -------------------------------
nltk==3.9.1
python-dateutil==2.9.0.post0

# -------------------------------
# Optional (for WhatsApp integration)
# -------------------------------
twilio==9.2.3
